+++ R/acid_base_equilibrium.R
@@ -38,10 +38,12 @@ solve_ph <- function(water, so4_dose = 0, na_dose = 0, ca_dose = 0, mg_dose = 0,
   return(round(phfinal, 2))
 }
 
-#' Add chemicals to water and apply acid/base equilibrium and ion mass balance
+#' @title Calculate new pH and ion balance after chemical addition
 #'
-#' \code{chemdose_ph} calculates the new pH, alkalinity, and ion balance of a water based on different chemical
-#' additions. The function takes an object of class "water" created by \code{\link{define_water}} and user-specified
+#' @description \code{chemdose_ph} calculates the new pH, alkalinity, and ion balance of a water based on different chemical
+#' additions.
+#'
+#' @details The function takes an object of class "water" created by \code{\link{define_water}} and user-specified
 #' chemical additions and returns a new object of class "water" with updated water quality.
 #' Units of all chemical additions are in mg/L as chemical (not as product).
 #'
@@ -60,6 +62,7 @@ solve_ph <- function(water, so4_dose = 0, na_dose = 0, ca_dose = 0, mg_dose = 0,
 #' @param caco3 Amount of calcium carbonate added (or removed) in mg/L: CaCO3 -> Ca + CO3
 #' @param caoh2 Amount of lime added in mg/L: Ca(OH)2 -> Ca + 2OH
 #' @param mgoh2  Amount of magneisum hydroxide added in mg/L: Mg(OH)2 -> Mg + 2OH
+#' @param cacl2 Amount of calcium chloride added in mg/L: CaCl2 -> Ca2+ + 2Cl-
 #' @param cl2 Amount of chlorine gas added in mg/L as Cl2: Cl2(g) + H2O -> HOCl + H + Cl
 #' @param naocl Amount of sodium hypochlorite added in mg/L as Cl2: NaOCl -> Na + OCl
 #' @param caocl2 Amount of calcium hypochlorite added in mg/L as Cl2: Ca(OCl)2 -> Ca + 2OCl
@@ -97,7 +100,7 @@ solve_ph <- function(water, so4_dose = 0, na_dose = 0, ca_dose = 0, mg_dose = 0,
 #' @export
 #'
 chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0, na2co3 = 0, nahco3 = 0, caco3 = 0, caoh2 = 0, mgoh2 = 0,
-                        cl2 = 0, naocl = 0, caocl2 = 0, co2 = 0,
+                        cacl2 = 0, cl2 = 0, naocl = 0, caocl2 = 0, co2 = 0,
                         alum = 0, fecl3 = 0, fe2so43 = 0,
                         softening_correction = FALSE) {
 
@@ -133,6 +136,9 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0, na2co3 =
   # Magnesium hydroxide (Mg(OH)2) dose
   mgoh2 = convert_units(mgoh2, "mgoh2")
 
+  # Calcium chloride (CaCl2) dose
+  cacl2 = convert_units(cacl2, "cacl2")
+
   # Chlorine gas (Cl2)
   cl2 = convert_units(cl2, "cl2")
 
@@ -162,7 +168,7 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0, na2co3 =
   dosed_water@na = water@na + na_dose
 
   # Total calcium
-  ca_dose = caoh2 + caocl2 / 2 + caco3
+  ca_dose = caoh2 + cacl2 + caocl2 / 2 + caco3
   dosed_water@ca = water@ca + ca_dose
 
   # Total magnesium
@@ -174,7 +180,7 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0, na2co3 =
   dosed_water@k = water@k + k_dose
 
   # Total chloride
-  cl_dose = hcl + cl2 + 3 * fecl3
+  cl_dose = hcl + cl2 + 2 * cacl2 + 3 * fecl3
   dosed_water@cl = water@cl + cl_dose
 
   # Total sulfate
@@ -251,9 +257,9 @@ chemdose_ph <- function(water, hcl = 0, h2so4 = 0, h3po4 = 0, naoh = 0, na2co3 =
 
 
 
-#' Calculate a desired chemical dose for a target pH
+#' @title Calculate a desired chemical dose for a target pH
 #'
-#' \code{solvedose_ph} calculates the required amount of a chemical to dose based on a target pH and existing water quality.
+#' @description \code{solvedose_ph} calculates the required amount of a chemical to dose based on a target pH and existing water quality.
 #' The function takes an object of class "water" created by \code{\link{define_water}}, and user-specified chemical and target pH
 #' and returns a numeric value for the required dose in mg/L.
 #'
@@ -331,9 +337,9 @@ solvedose_ph <- function(water, target_ph, chemical) {
 }
 
 
-#' Calculate a desired chemical dose for a target alkalinity
+#' @title Calculate a desired chemical dose for a target alkalinity
 #'
-#' This function calculates the required amount of a chemical to dose based on a target alkalinity and existing water quality.
+#' @description This function calculates the required amount of a chemical to dose based on a target alkalinity and existing water quality.
 #' Returns numeric value for dose in mg/L. Uses uniroot on the chemdose_ph function.
 #'
 #' @param water Source water of class "water" created by \code{\link{define_water}}
@@ -399,9 +405,9 @@ solvedose_alk <- function(water, target_alk, chemical) {
 }
 
 
-#' Determine blended water quality from multiple waters based on mass balance and acid/base equilibrium
+#' @title Determine blended water quality from multiple waters based on mass balance and acid/base equilibrium
 #'
-#' This function takes a vector of waters defined by \code{\link{define_water}} and a vector of ratios and outputs a new water object with updated ion balance and pH.
+#' @description This function takes a vector of waters defined by \code{\link{define_water}} and a vector of ratios and outputs a new water object with updated ion balance and pH.
 #'
 #' @param waters Vector of source waters created by \code{\link{define_water}}
 #' @param ratios Vector of ratios in the same order as waters. (Blend ratios must sum to 1)